Moved indicator to advanced settings page, refactored
