+++ src/generic_ui/locales/en/messages.json
@@ -656,11 +656,11 @@
     "message": "Restart to update"
   },
   "portControlSupported": {
-    "description": "NAT-PMP, PCP, or UPnP is supported on the router",
+    "description": "NAT-PMP, PCP, or UPnP is supported in the router",
     "message": "Port control supported"
   },
   "portControlNotSupported": {
-    "description": "NAT-PMP, PCP, and UPnP are not supported on the router",
+    "description": "NAT-PMP, PCP, and UPnP are not supported in the router",
     "message": "Port control not supported"
   }
 }
