+++ src/generic_ui/polymer/advanced-settings.html
@@ -88,6 +88,28 @@
     .error {
       color: rgb(200, 112, 112);
     }
+    #portControlBox core-icon[icon='help'],
+    #portControlBox core-icon[icon='refresh'] {
+      height: 15px;
+      color: grey;
+      opacity: 0.6;
+      margin-bottom: 2px;
+      margin-right: 0px;
+      -webkit-transition: all 0.23s !important;
+      -moz-transition: all 0.23s !important;
+      transition: all 0.23s !important;
+    }
+    #portControlBox core-icon[icon='help']:hover,
+    #portControlBox core-icon[icon='refresh']:hover {
+      opacity: 1;
+    }
+    #portControlBox span, #portControlBox uproxy-link,
+    #portControlBox uproxy-faq-link {
+      display: table-cell;
+    }
+    #portControlBox span {
+      color: rgb(112, 112, 112);
+    }
     </style>
 
     <!-- Listen for the 'open-advanced-settings' event, which is fired by the settings page. -->
@@ -116,6 +138,21 @@
           <p id='failedKeyValueSetAdvancedSettings' class='advancedSettingsText error' hidden?='{{ status !== StatusState.KEY_VALUE_ERROR }}'>
             {{ 'settingsJsonError' | $$ }}
           </p>
+
+          <div id='portControlBox'>
+            <span class='label' hidden?='{{ !ui.portControlSupport }}'>
+              {{ 'portControlSupported' | $$ }}
+            </span>
+            <span class='label' hidden?='{{ ui.portControlSupport }}'>
+              {{ 'portControlNotSupported' | $$ }}
+            </span>
+            <uproxy-link on-tap='{{ refreshPortControl }}' role='button'>
+              <core-icon icon='refresh'></core-icon>
+            </uproxy-link>
+            <uproxy-faq-link anchor='whatIsPortControl'>
+              <core-icon icon='help'></core-icon>
+            </uproxy-link>
+          </div>
         </div> <!-- end of form container -->
         <div id='setAdvancedSettings' on-tap='{{ setAdvancedSettings }}'>
           {{ 'set' | $$ }}
